rf: Rename dataset-alert-draft.tsx to dataset-alert.tsx
