+++ packages/openneuro-app/src/scripts/dataset/draft-container.tsx
@@ -13,7 +13,7 @@ import {
   hasEditPermissions,
 } from "../authentication/profile"
 import { useCookies } from "react-cookie"
-import { DatasetAlertDraft } from "./fragments/dataset-alert-draft"
+import { DatasetAlertDraft } from "./fragments/dataset-alert"
 import { CloneDropdown } from "./components/CloneDropdown"
 import { DatasetGitAccess } from "./components/DatasetGitAccess"
 import { DatasetHeader } from "./components/DatasetHeader"
